Renamed Id / roomId for clarity
