+++ backend/resolvers/resolvers.root.js
@@ -1,5 +1,5 @@
-const UsersQuery = require("./users/users.query");
-const UsersResolver = require("./users/users.resolver");
+const UsersQuery = require("./voters/voters.query");
+const UsersResolver = require("./voters/voters.resolver");
 const RoomsQuery = require("./rooms/rooms.query");
 const RoomsResolver = require("./rooms/rooms.resolver");
