+++ frontend/pages/development/name-generator.js
@@ -1,19 +1,14 @@
+import NameGenerator from "../../components/userNames";
+
 import Link from "next/link";
 import styled from "styled-components";
-import { COLORS } from "../../styles/colors";
+import Layout from "../../components/layout";
 
-import Timer from "../../components/timer";
-
-export default function VotingRoom() {
+export default function LandingPage() {
   return (
     <Container>
-      <Header>Voting Page</Header>
-
-      <Timer onTimeIsUp={(message) => alert(message)} />
-
-      <Description>
-        This page will be where the voting itself takes place
-      </Description>
+      <Header>Name Generator</Header>
+      <NameGenerator />
       <Link href="/">
         <Button>Home</Button>
       </Link>
@@ -23,11 +18,7 @@ export default function VotingRoom() {
 
 const Container = styled.div`
   /* background-color: ${COLORS.MAIN}; */
-  background: linear-gradient(
-    to left top,
-    ${COLORS.SHADES.WHITE} 50%,
-    ${COLORS.MAIN} 50%
-  );
+  background: linear-gradient(to left top, #fff 50%, #eb5e28 50%);
   height: 100vh;
   display: flex;
   align-items: center;
@@ -42,11 +33,6 @@ const Header = styled.h1`
   font-size: 3rem;
 `;
 
-const Description = styled.p`
-  color: #293241;
-  margin-left: 1rem;
-`;
-
 const Button = styled.button`
   margin-top: 15px;
   padding: 8px;
@@ -57,8 +43,8 @@ const Button = styled.button`
   cursor: pointer;
 
   &:active {
-    background: ${COLORS.SHADES.GREY};
-    box-shadow: inset 0px 0px 5px ${COLORS.SHADES.DARKGREY};
+    background: #e5e5e5;
+    box-shadow: inset 0px 0px 5px #c1c1c1;
     outline: none;
     transform: scale(0.9);
   }
